chore: add version and plugin dependencies files, and generate pubspec.lock
- Created version file with version 3.38.4.
- Added flutter-plugins-dependencies file with plugin details for iOS, Android, macOS, Linux, Windows, and web.
- Generated pubspec.lock file with dependencies including async, dio, flutter_secure_storage, shared_preferences, and others.

+++ jan_sampark_mobile/lib/core/utils/extensions.dart
@@ -142,7 +142,7 @@ extension IntExtensions on int {
 }
 
 extension DoubleExtensions on double {
-  double get asPercent => clamp(0.0, 100.0);
+  double get asPercent => clamp(0.0, 100.0).toDouble();
   SizedBox get h => SizedBox(height: this);
   SizedBox get w => SizedBox(width: this);
 }
